requirements.txt
torch>=1.9.0
numpy>=1.19.2
pandas>=1.1.3
librosa>=0.8.0
scipy>=1.5.2
pyyaml>=5.3.1
matplotlib>=3.3.2
tqdm>=4.50.2
